Move Location to geo package
