+++ core/entities/src/commonMain/kotlin/geo/Location.kt
@@ -2,9 +2,7 @@
  * Copyright (c) 2024 Pointyware. Use of this software is governed by the GPL-3.0 license.
  */
 
-package org.pointyware.xyz.core.entities.ride
-
-import org.pointyware.xyz.core.entities.geo.LatLong
+package org.pointyware.xyz.core.entities.geo
 
 /**
  * Associates useful information with [LatLong] coordinates.
