+++ feature/ride/src/commonMain/kotlin/viewmodels/RideUiState.kt
@@ -5,7 +5,7 @@
 package org.pointyware.xyz.feature.ride.viewmodels
 
 import org.pointyware.xyz.core.entities.business.Currency
-import org.pointyware.xyz.core.entities.ride.Location
+import org.pointyware.xyz.core.entities.geo.Location
 import org.pointyware.xyz.core.entities.ride.Route
 import org.pointyware.xyz.core.viewmodels.BriefProfileUiState
